+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2025-11-24
+
+### Added
+
+- added worker start behaviour (link, monitor, nolink)
+- returnes the expected message size in the decoding function in case of an incompleted message
+
+### Fixed
+
+- increased tcp buffer size
+- set active-once mode to recv tcp data
+- changed msgpack options
+- deleted pending request on timeout
+- reset worker state on disconnection
+
+### Misc
+
+- moved get_error function to tnt_proto module
+- updated README
+
 ## [1.0.1] - 2025-10-13
 
 ### Added
@@ -35,5 +55,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Request queue management
 - MessagePack serialization via the `msgpack` library
 
+[1.0.2]: https://github.com/yak-zuk-zop/tnt/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/yak-zuk-zop/tnt/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/yak-zuk-zop/tnt/releases/tag/v1.0.0